feat(sdk): add dual-ISP depth compute support for ADSD3500
Implement non-ISP depth computation path with lens scatter compensation for ADSD3500 dual-ISP modes.
Adds raw CCB reading (195KB), expands calDataLength to uint32_t, and distinguishes V4L2 raw bypass from
ToFi processing bypass. Enables mode 1 to output 1024×1024 depth+AB from 2048×4608 raw sensor data.

+++ sdk/include/aditof/depth_sensor_interface.h
@@ -278,7 +278,7 @@ class DepthSensorInterface {
     virtual aditof::Status initTargetDepthCompute(uint8_t *iniFile,
                                                   uint16_t iniFileLength,
                                                   uint8_t *calData,
-                                                  uint16_t calDataLength) = 0;
+                                                  uint32_t calDataLength) = 0;
 
     /**
      * @brief Get ini parameters for Depth Compute library
