+++ sdk/src/connections/network/network_depth_sensor.h
@@ -86,7 +86,7 @@ class NetworkDepthSensor : public aditof::DepthSensorInterface {
                                                int &imagerStatus) override;
     virtual aditof::Status
     initTargetDepthCompute(uint8_t *iniFile, uint16_t iniFileLength,
-                           uint8_t *calData, uint16_t calDataLength) override;
+                           uint8_t *calData, uint32_t calDataLength) override;
     virtual aditof::Status
     getDepthComputeParams(std::map<std::string, std::string> &params) override;
