+++ sdk/src/connections/offline/offline_depth_sensor.cpp
@@ -568,7 +568,7 @@ aditof::Status OfflineDepthSensor::adsd3500_get_status(int &chipStatus,
  */
 aditof::Status OfflineDepthSensor::initTargetDepthCompute(
     uint8_t *iniFile, uint16_t iniFileLength, uint8_t *calData,
-    uint16_t calDataLength) {
+    uint32_t calDataLength) {
 
     using namespace aditof;
